requirement.txt
streamlit
pillow
paddleocr
torch
ultralytics
numpy
